backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
google-genai==0.7.0
pydantic==2.10.6
python-dotenv==1.0.1
google-cloud-texttospeech==2.18.0
pytest==8.3.4
httpx==0.27.2
